Support TouchableNativeFeedback for Android (#86)
Resolves issues 85, 84, 57.

To achieve proper styling, render method now conditionally renders button for iOS and Android. Added new prop 'androidBackground' which should be one of TouchableNativeFeedback.ThemeAttributeBackgroundPropType. Since TouchableNativeFeedback doesn't have proper styling, it is wrapped in a View which receives containerStyle prop. However, it doesn't work well with padding, so another View added as a nested component, and containerStyle.padding is being set to this nested View.